refactor auth setup
