+++ cli/src/installers/nextAuth.ts
@@ -6,8 +6,9 @@ import { PKG_ROOT } from "~/consts.js";
 import { addPackageDependency } from "~/utils/addPackageDependency.js";
 
 export const nextAuthInstaller: Installer = ({ projectDir, packages }) => {
+  const usingPrisma = packages?.prisma.inUse;
   const deps: AvailableDependencies[] = ["next-auth"];
-  if (packages?.prisma.inUse) deps.push("@next-auth/prisma-adapter");
+  if (usingPrisma) deps.push("@next-auth/prisma-adapter");
 
   addPackageDependency({
     projectDir,
@@ -17,23 +18,17 @@ export const nextAuthInstaller: Installer = ({ projectDir, packages }) => {
 
   const extrasDir = path.join(PKG_ROOT, "template/extras");
 
-  const apiHandlerSrc = path.join(
+  const apiHandlerFile = "src/pages/api/auth/[...nextauth].ts";
+  const apiHandlerSrc = path.join(extrasDir, apiHandlerFile);
+  const apiHandlerDest = path.join(projectDir, apiHandlerFile);
+
+  const authConfigSrc = path.join(
     extrasDir,
-    "src/pages/api/auth/[...nextauth]",
-    packages?.prisma.inUse ? "with-prisma.ts" : "base.ts",
+    "src/server",
+    usingPrisma ? "with-prisma.ts" : "base.ts",
   );
-  const apiHandlerDest = path.join(
-    projectDir,
-    "src/pages/api/auth/[...nextauth].ts",
-  );
-
-  const getServerAuthSessionSrc = path.join(extrasDir, "src/server/auth.ts");
-  const getServerAuthSessionDest = path.join(projectDir, "src/server/auth.ts");
-
-  const nextAuthDTSSrc = path.join(extrasDir, "src/types/next-auth.d.ts");
-  const nextAuthDTSDest = path.join(projectDir, "src/types/next-auth.d.ts");
+  const authConfigDest = path.join(projectDir, "src/server/auth.ts");
 
   fs.copySync(apiHandlerSrc, apiHandlerDest);
-  fs.copySync(getServerAuthSessionSrc, getServerAuthSessionDest);
-  fs.copySync(nextAuthDTSSrc, nextAuthDTSDest);
+  fs.copySync(authConfigSrc, authConfigDest);
 };
