+++ cli/template/extras/src/pages/api/auth/[...nextauth].ts
@@ -1,0 +1,4 @@
+import NextAuth from "next-auth";
+import { authOptions } from "../../../server/auth";
+
+export default NextAuth(authOptions);
