--- cli/template/extras/src/server/auth.ts
@@ -1,17 +1,0 @@
-import { type GetServerSidePropsContext } from "next";
-import { getServerSession } from "next-auth";
-
-import { authOptions } from "../pages/api/auth/[...nextauth]";
-
-/**
- * Wrapper for getServerSession so that you don't need
- * to import the authOptions in every file.
- * @see https://next-auth.js.org/configuration/nextjs
- */
-
-export const getServerAuthSession = async (ctx: {
-  req: GetServerSidePropsContext["req"];
-  res: GetServerSidePropsContext["res"];
-}) => {
-  return await getServerSession(ctx.req, ctx.res, authOptions);
-};
